1.2 Update
This commit brings the 1.2 version of PSMultiserver on GitHub.

This will mark the introduction of many features and fixes to booth, the server software and it's features.

Most noticable is the introduction of Horizon into the project (the Jumpsuit fork), which has been ported to net 6.0, as well as the VEEMEE api, built from scratch in c#.