Multiserver 2 Project update.
A completely re-designed server suite, with the aim of being much faster, and much more scalable.

Can bring up-to a 70% reduced memory usage under stress. And a performance in HTTP tests that can result in a 2 times increase in performance.

There are some many additions in this new project, that reading full code is best than making a list of changes at that point.