[NetworkLibrary] - Implemented multi-threaded Json directory maker + minor code cleanup in that same area.
Also sets a thread sleep in the while loop of ClientProcess in the http server, no need to process the requests that fast if someone (like a bot) spam 404 requests.

Additionally adjusts slightly the main loops of HTTP and SVO server queues to be more secure in case of a reboot/hard crash.