[ApacheNet/WatsonWebserver/NetworkLibrary] - Fixes stream related events memory leak, further improves the html directory report, fixes some remaining bugs.
This commit removes the stream error inhibition which caused stream-tied events (such as ffmpeg streaming) to remain in memory even after the client closed the connection.

This also further improves the directory html report, with a built-in video transcoder for incompatible video formats.

Also fixes some bugs in the apache client handler, and moves the process utils class to the NetworkLibrary.