[CastleLibrary/QuazalServer] - Bouncy Castle upstream + Quazal semi code refactor (PC support).
This commit updates the bouncy castle revision to the latest to this date (+ the custom LIBSECURE class added) and refactor/improve the Quazal server to properly implement PC support. This also updates the WebAPIs to handle Driver San Fransisco PC version web-requests.

The list of Quazal changes are as follows:

- Memory leak fixes related to compression.
- Refactoring of the RMCId calculation, this is now separated per game classes to avoids duplicates.
- Temporarly removes the Quazal Database for console players, in favor of a in-memory account fetching and live-storage.
- Implements DFS PC game class.
- Inserts 2 generic accounts in the Uplay database by default, to allow guests players to connect, and to test the server.
- Minor code changes to unify all the former changes.