[MutiSocks/NetworkLibrary/Horizon/SVO] - Moves Horizon DateTime utils to the NetworkLibrary + fixup DirtySocks game list player ordering handling.
This commit uses the awesome Horizon date time utils to help improving the DirtySocks game player ordering issues there was (a player could leave and confuse the order of the listing somehow).

What it does is to then, modify the rest of the code accordingly to use this newer unified class.