[MultiSocks/MitmDNS/DOH/Horizon] - Optimize more the DNS configs fetching + DirtySocks fixes, and a slight using fixup in World.cs.
The original DNS fetching had hugely wasted speed potential, that was fixed up by using concurrent storage types when applicable.

This commit also simplifies the user storage in the DirtySocks UsersCollection, as the game itself is supposed to have it's own player index system. Speaking of which, this system is now properly emulated, properly fixing the player colors in BP while in a game session. This also fixes tha MA report in the +who packet.

This also adjust some usings for the World.cs file in the Horizon project.