[Horizon/HorizonService/BackendService] - Upstream update + partial code cleanup.
This commit on Horizon based servers introduced the Ping ban check from upstream, fixes some code logger formatting, removes the non-working DME encryption setting, and updates the Party class to the same model as the Game one (NEEDS TESTING).

In the backend, it removes the threading code for one of the array copy methods to not preasue the limited thread pool too much.